refactor analysis slightly, add docs for analysis to sphinx, restructure notebook, started CLI for analysis
