+++ be_scan/analysis/__init__.py
@@ -1,3 +1,6 @@
 
-from be_scan.analysis._count_reads import count_reads
+from be_scan.analysis.count_reads import count_reads
+from be_scan.analysis.merge_and_norm import merge_and_norm
+from be_scan.analysis.average_reps import average_reps
+
 from be_scan.analysis._sanger import validate_cloning
