+++ be_scan/analysis/_count_reads.py
@@ -22,328 +22,7 @@ from sklearn.preprocessing import StandardScaler
 from sklearn.decomposition import PCA
 import gzip
 
-#%% count_reads() - count sgRNA reads in FASTQ (adapted from count_spacers)
 
-def count_reads(in_fastq, in_ref, KEY_INTERVAL=(10,80),
-                KEY='CGAAACACC', KEY_REV='GTTTTAGA', out_counts='counts.csv',
-                out_np='np_counts.csv', out_stats='stats.txt', dont_trim_G=False):
-    """Count the reads in a FASTQ file and assign them to a reference sgRNA set.
-
-    Given a set of sgRNA sequences and a FASTQ file, count the reads in the
-    FASTQ, assign the reads to sgRNAs, and export the counts to a csv file `out_counts`. All
-    sgRNA sequences not found in the reference file (non-perfect matches) are
-    written to a separate csv file `out_np`.
-
-    Parameters
-    ----------
-    in_fastq : str or path
-        String or path to the FASTQ file to be processed.
-    in_ref : str or path
-        String or path to the reference file. in_ref must have column headers,
-        with 'sgRNA_seq' as the header for the column with the sgRNA sequences.
-    KEY_INTERVAL : tuple, default (10,80)
-        Tuple of (KEY_START, KEY_END) that defines the KEY_REGION. Denotes the
-        substring within the read to search for the KEY.
-    KEY : str, default 'CGAAACACC'
-        Sequence that is expected upstream of the spacer sequence. The default
-        is the end of the hU6 promoter.
-    KEY_REV : str, default 'GTTTTAGA'
-        Sequence that is expected downstream of the spacer sequence. The
-        default is the start of the sgRNA scaffold sequence.
-    out_counts : str or path, default 'counts.csv'
-        String or path for the output csv file with perfect sgRNA matches.
-    out_np : str or path, default 'np_counts.csv'
-        String or path for the output csv file with non-perfect sgRNA matches.
-    out_stats : str or path, default 'stats.txt'
-        String or path for the output txt file with the read counting statistics.
-    dont_trim_G : bool, default False
-        Whether to trim the first G from 21-nt sgRNA sequences to make them 20-nt.
-    """
-
-    # STEP 1A: OPEN INPUT FILES FOR PROCESSING, CHECK FOR REQUIRED FORMATTING
-    # look for 'sgRNA_seq' column, raise Exception if missing
-    df_ref = pd.read_csv(in_ref, header=0) # explicit header = first row
-    if 'sgRNA_seq' not in df_ref.columns.tolist():
-        raise Exception('in_ref is missing column: sgRNA_seq')
-    # look for other cols, raise Warning if suggested cols are missing
-    list_headcols = ['sgRNA_ID', 'sgRNA_seq', 'Gene', 'cut_site_AA', 'Domain']
-    if not all(col in df_ref.columns.tolist() for col in list_headcols):
-        list_miss = [col for col in list_headcols if col not in df_ref.columns.tolist()]
-        warnings.warn('Warning! in_ref is missing column(s) for downstream functions: ' + str(list_miss))
-    # try opening input FASTQ, raise Exception if not possible
-    if in_fastq.endswith('.gz'):
-        handle = gzip.open(in_fastq, 'rt')
-    else:
-        handle = open(in_fastq, 'rt')
-
-    # STEP 1B: SET UP VARIABLES FOR SCRIPT
-    # make dictionary to hold sgRNA counts - sgRNA_seq, count as k,v
-    dict_perfects = {sgRNA:0 for sgRNA in df_ref['sgRNA_seq']}
-    list_np = [] # placeholder list for non-perfect matches
-    num_reads = 0 # total number of reads processed
-    num_perfect_matches = 0 # count of reads with a perfect match to library
-    num_np_matches = 0 # count of reads without a perfect match to library
-    num_nokey = 0 # count of reads where key was not found
-    num_badlength = 0 # count of sgRNA reads that aren't 20bp
-    KEY_START, KEY_END = KEY_INTERVAL[0], KEY_INTERVAL[1] # set the key interval
-
-    # STEP 2: PROCESS FASTQ FILE READS AND ADD COUNTS TO DICT
-    while True: # contains the seq and Qscore etc.
-        read = handle.readline()
-        if not read: # end of file
-            break
-        elif read.startswith("@"): # if line is a read header
-            read = handle.readline() # next line is the actual sequence
-        else:
-            continue
-        num_reads += 1
-        read_sequence = str.upper(str(read))
-        key_region = read_sequence[KEY_START:KEY_END]
-        key_index = key_region.find(KEY)
-        key_rev_index = key_region.find(KEY_REV)
-        if key_index < 0 or key_rev_index <= key_index: # if keys not found
-            num_nokey += 1
-            continue
-        start_index = key_index + KEY_START + len(KEY)
-        end_index = key_rev_index + KEY_START
-        guide = read_sequence[start_index:end_index]
-        if not dont_trim_G:
-            if guide.startswith('G') and len(guide) == 21:
-                guide = guide[1:]
-        if len(guide) != 20:
-            num_badlength += 1
-            continue
-        if guide in dict_perfects:
-            dict_perfects[guide] += 1
-            num_perfect_matches += 1
-        else:
-            num_np_matches += 1
-            list_np.append(guide)
-    handle.close()
-
-    # STEP 3: SORT DICTIONARIES AND GENERATE OUTPUT FILES
-    # sort perf matches (A-Z) with guides,counts as k,v and output to csv
-    df_perfects = pd.DataFrame(data=dict_perfects.items(), columns=['sgRNA_seq', 'reads'])
-    df_perfects.sort_values(by='sgRNA_seq', inplace=True)
-    df_perfects.to_csv(out_counts, index=False, header=False)
-    # now sort non-perfect matches by frequency and output to csv
-    dict_np = Counter(list_np) # use Counter to tally up np matches
-    df_npmatches = pd.DataFrame(data=dict_np.items(), columns=['sgRNA_seq', 'reads'])
-    df_npmatches.sort_values(by='reads', ascending=False, inplace=True)
-    df_npmatches.to_csv(out_np, index=False)
-
-    # STEP 4: CALCULATE STATS AND GENERATE STAT OUTPUT FILE
-    # percentage of guides that matched perfectly
-    pct_perfmatch = round(num_perfect_matches/float(num_perfect_matches + num_np_matches) * 100, 1)
-    # percentage of undetected guides (no read counts)
-    guides_with_reads = np.count_nonzero(list(dict_perfects.values()))
-    guides_no_reads = len(dict_perfects) - guides_with_reads
-    pct_no_reads = round(guides_no_reads/float(len(dict_perfects.values())) * 100, 1)
-    # skew ratio of top 10% to bottom 10% of guide counts
-    top_10 = np.percentile(list(dict_perfects.values()), 90)
-    bottom_10 = np.percentile(list(dict_perfects.values()), 10)
-    if top_10 != 0 and bottom_10 != 0:
-        skew_ratio = top_10/bottom_10
-    else:
-        skew_ratio = 'Not enough perfect matches to determine skew ratio'
-    # calculate the read coverage (reads processed / sgRNAs in library)
-    num_guides = df_ref['sgRNA_seq'].shape[0]
-    coverage = round(num_reads / num_guides, 1)
-    # calculate the number of unmapped reads (num_nokey / total_reads)
-    pct_unmapped = round((num_nokey / num_reads) * 100, 2)
-
-    # write analysis statistics to statfile
-    with open(out_stats, 'w') as statfile:
-        statfile.write('Number of reads processed: ' + str(num_reads) + '\n')
-        statfile.write('Number of reads where key was not found: ' + str(num_nokey) + '\n')
-        statfile.write('Number of reads where length was not 20bp: ' + str(num_badlength) + '\n')
-        statfile.write('Number of perfect guide matches: ' + str(num_perfect_matches) + '\n')
-        statfile.write('Number of nonperfect guide matches: ' + str(num_np_matches) + '\n')
-        statfile.write('Number of undetected guides: ' + str(guides_no_reads) + '\n')
-        statfile.write('Percentage of unmapped reads (key not found): ' + str(pct_unmapped) + '\n')
-        statfile.write('Percentage of guides that matched perfectly: ' + str(pct_perfmatch) + '\n')
-        statfile.write('Percentage of undetected guides: ' + str(pct_no_reads) + '\n')
-        statfile.write('Skew ratio of top 10% to bottom 10%: ' + str(skew_ratio) + '\n')
-        statfile.write('Read coverage: ' + str(coverage))
-        statfile.close()
-
-    print(str(in_fastq) + ' processed')
-    return
-
-#%% merge_and_norm() - merge sample read counts and log2/t0 normalize
-
-def merge_and_norm(dict_counts, in_ref, t0='t0', dir_counts='', save='all',
-                   out_folder='', out_reads='agg_reads.csv', out_log2='agg_log2.csv',
-                   out_t0='agg_t0_reps.csv', return_df=None):
-    """
-    Aggregate raw counts and perform log2-transform and t0 normalization.
-
-    For a given set of samples and their raw read count files from count_reads,
-    aggregate them into a single dataframe, normalize counts to reads per million,
-    perform log2 transform, and then normalize to t0. The aggregated raw reads,
-    log2 transformed values, and t0 normalized values can be saved to csv.
-
-    Parameters
-    ----------
-    dict_counts : dict in format {'sample name': 'file name'}
-        Dictionary to map sample names (key) to read count file names (value),
-        as key,value (e.g. {'KN-0': 'KN-0_counts.csv'}). Must include the t0
-        sample in the dict. 
-    in_ref : str or path
-        String or path to the reference file. in_ref must have column headers,
-        with 'sgRNA_seq' as the header for the column with the sgRNA sequences.
-    t0 : str, default 't0'
-        Name of the t0 sample in dict_counts. If you have multiple t0 samples
-        (e.g. paired t0s for specific samples), then you will need to run this
-        function separately for each set of samples with their appropriate t0.
-    dir_counts : str, default ''
-        Name of the subfolder to find the read count csv files. The default is
-        the current working directory.
-    save : {'all', None, ['reads', 'log2', 't0']}, default 'all'
-        Choose files for export to csv. The default is 'all', which is the
-        aggregated read counts ('reads'), log2 normalized values ('log2'), and
-        t0 normalized values ('t0'). You may also enter any combination of
-        'reads', 'log2', 't0') as a list of strings to choose which ones to
-        save ('all' is equivalent to a list of all three). None will not export
-        any files to csv.
-    out_folder : str, default ''
-        Name of the subfolder to save output files. The default is the current
-        working directory.
-    out_reads : str, default 'agg_reads.csv'
-        Name of the aggregated raw reads csv output file.
-    out_log2 : str, default 'agg_log2.csv'
-        Name of the aggregated log2 normalized values csv output file.
-    out_t0 : str, default 'agg_t0_reps.csv'
-        Name of the aggregated t0 normalized values csv output file.
-    return_df : {None, 'reads', 'log2', 't0'}, default None
-        Whether to return a dataframe at function end. The default is None,
-        which returns nothing. However, you can return the reads, log2 norm,
-        or t0 norm values by calling 'reads', 'log2', or't0', respectively.
-    """
-
-    # import reference file, define variables, check for requirements
-    path = Path.cwd()
-    inpath = path / dir_counts
-    df_ref = pd.read_csv(in_ref)
-    if 'sgRNA_seq' not in df_ref.columns.tolist():
-        raise Exception('in_ref is missing column: sgRNA_seq')
-    if t0 not in dict_counts.keys():
-        raise Exception ('dict_counts is missing the t0 sample')
-    # rearrange the dict samples to place t0 first (for order purposes later)
-    list_samples = [t0] + [samp for samp in dict_counts.keys() if samp != t0]
-    # aggregate read counts from all samples into df_rawreads
-    # also perform log2 norm (brian/broad method; log2(rpm + 1 / total reads))
-    df_reads, df_log2, df_t0 = df_ref.copy(), df_ref.copy(), df_ref.copy()
-    for sample in list_samples:
-        df_temp = pd.read_csv(inpath / dict_counts[sample], names=['sgRNA_seq', sample])
-        # aggregating raw reads
-        df_reads = pd.merge(df_reads, df_temp, on='sgRNA_seq')
-        # log2 normalization
-        total_reads = df_reads[sample].sum()
-        df_log2[sample] = df_reads[sample].apply(lambda x: np.log2((x * 1000000 / total_reads) + 1))
-        # t0 normalization
-        df_t0[sample] = df_log2[sample].sub(df_log2[t0])
-    # drop the t0 column since it will be 0
-    df_t0.drop(columns=t0, inplace=True)
-
-    # export files and return dataframes if necessary
-    outpath = path / out_folder
-    Path.mkdir(outpath, exist_ok=True)
-    # dictionary to map kws to dfs and output file names
-    dict_df = {'reads': (df_reads, out_reads), 'log2': (df_log2, out_log2), 't0': (df_t0, out_t0)}
-    # determine which files to export
-    if save == 'all':
-        save = ['reads','log2','t0']
-    if isinstance(save, list):
-        for key in save:
-            dict_df[key][0].to_csv(outpath / dict_df[key][1], index=False)
-    elif save is None:
-        pass
-    else:
-        warnings.warn('Invalid value for save. No files exported')
-    # determine df to return
-    print('Merge and normalize completed')
-    if return_df in dict_df.keys():
-        return dict_df[return_df][0]
-    elif return_df is None:
-        return
-    else:
-        print('Invalid value for return_df. No dataframe returned')
-        return
-
-#%% average_reps() - average replicates by condition
-
-def average_reps(dict_conds, in_lfc, in_ref, save=True, out_folder='',
-                 out_conds='agg_t0_conds.csv', return_df=False):
-    """
-    For a set of conditions, average the replicates and export it to csv.
-
-    Averages the replicates for each condition (e.g. treatment, control) and
-    exports the csv file with averaged replicate values for each condition.
-    Note that this function is generally used for averaging t0 normalized
-    replicate values, but can be used on any set of replicate values
-    (e.g. log2 fc values or even raw reads).
-
-    Parameters
-    ----------
-    dict_conds : dict in format {'replicate': 'condition'}
-        Dictionary to map replicates (key) to conditions (value). Replicates
-        must match the column headers in the in_lfc file (i.e. sample names),
-        otherwise they will be discarded before performing the averaging.
-        Example: {'KN-1': 'unsorted', 'KN-2': 'unsorted', 'KN-3': 'sorted'}
-    in_lfc : str or path
-        String or path to the individual replicate values csv file. The column
-        headers for the replicate values must match the keys in dict_conds
-    in_ref : str or path
-        String or path to the reference file. in_ref must have column headers,
-        with 'sgRNA_seq' as the header for the column with the sgRNA sequences.
-    save : bool, default True
-        Whether to save the averaged replicate values as a csv file.
-    out_folder : str, default ''
-        Name of the subfolder to save output files. The default is the current
-        working directory.
-    out_conds : str, default 'agg_t0_conds.csv'
-        Name of the averaged replicate values csv output file.
-    return_df : bool, default False
-        Whether to return the averaged reps dataframe The default is False.
-    """
-
-    # import files, define variables, check for requirements, etc.
-    path = Path.cwd()
-    df_lfc = pd.read_csv(in_lfc)
-    # make df to map replicates to condition
-    df_map = pd.DataFrame(data=dict_conds.items(), columns=['rep','condition'])
-    # check to make sure replicates are in the input file
-    list_reps = df_map['rep'].tolist()
-    if not all(rep in list_reps for rep in df_lfc.columns.tolist()):
-        list_miss = [rep for rep in list_reps if rep not in df_lfc.columns.tolist()]
-        # remove missing replicates from df_map and raise Warning
-        df_map = df_map.loc[~df_map['rep'].isin(list_miss)]
-        warnings.warn('in_lfc is missing replicates (removed from analysis): ' + str(list_miss))
-
-    # generate df to hold the averaged replicates per condition
-    df_conds = pd.read_csv(in_ref)
-    for cond in df_map['condition'].unique().tolist():
-        # for each unique condition, find the reps
-        reps = df_map.loc[df_map['condition'] == cond]['rep'].tolist()
-        # skip averaging for single replicates (otherwise breaks script)
-        if len(reps) > 1:
-            df_conds[cond] = df_lfc[reps].mean(axis=1)
-        elif len(reps) == 1:
-            df_conds[cond] = df_lfc[reps]
-        else:
-            raise Exception('Error! Replicate number not valid')
-
-    # export files and return dataframes if necessary
-    if save:
-        outpath = path / out_folder
-        Path.mkdir(outpath, exist_ok=True)
-        df_conds.to_csv(outpath / out_conds, index=False)
-    print('Average reps completed')
-    if return_df:
-        return df_conds
-    else:
-        return
 
 #%% compare_conds() - compare treatment vs. control to calculate enrichments
