+++ docs/source/plot.md
@@ -1,10 +1,14 @@
 # Default Plot Functions
 
+## plot_scatterplot()
+
 This function takes in a dataframe from count_reads, performs normalization, and then plots the data for each condition to reveal which guides are enriched {py:func}`be_scan.plot.plot_scatterplot`
 
 ```{eval-rst}
 .. autofunction:: be_scan.plot.plot_scatterplot
 ```
+
+## plot_corr_scatterplot()
 
 This function takes in a dataframe from count_reads, and plots a heatmap showing correlation between all listed conditions {py:func}`be_scan.plot.plot_corr_scatterplot`
 
@@ -12,11 +16,15 @@ This function takes in a dataframe from count_reads, and plots a heatmap showing
 .. autofunction:: be_scan.plot.plot_corr_scatterplot
 ```
 
+## plot_corr_heatmap()
+
 This function takes in a dataframe from count_reads, and plots a heatmap showing correlation between all listed conditions {py:func}`be_scan.plot.plot_corr_heatmap`
 
 ```{eval-rst}
 .. autofunction:: be_scan.plot.plot_corr_heatmap
 ```
+
+## plot_boxes()
 
 This function takes in a dataframe from count_reads, and plots a heatmap showing correlation between all listed conditions {py:func}`be_scan.plot.plot_boxes`
